[♻️]refact: Controller에게 위임하는 형태로 수정
